refactor: рефакторинг страницы MainPage по замечаниям
